% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{A function to add numbers}
\usage{
add(a, b)
}
\arguments{
\item{x}{A number.}

\item{y}{A number to add to x.}
}
\value{
A number that is the sum of a and b.
}
\description{
A function to add numbers
}
