% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_to_F.R
\name{C_to_F}
\alias{C_to_F}
\title{Celsius to Fahrenheit}
\usage{
C_to_F(x)
}
\arguments{
\item{x}{Numeric vector of temperatures in Celsius}
}
\value{
A numeric vector of temperatures in Fahrenheit
}
\description{
\code{C_to_F} converts the temperature units in Celsius to Fahrenheit
}
\examples{

C_to_F(32) # Converts from 32 C to Fahrenheit
C_to_F(44) # Converts from 44 C  to Fahrenheit

}
\seealso{
C_to_K, F_to_K
}
