% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_to_K.R
\name{C_to_K}
\alias{C_to_K}
\title{Celsius to Kelvin}
\usage{
C_to_K(x)
}
\arguments{
\item{x}{Numeric vector of temperatures in Celsius}
}
\value{
A numeric vector of temperatures in Kelvin
}
\description{
\code{C_to_K} converts the temperature units in Celsius to Kelvin
}
\examples{

C_to_K(32) # Converts from 32 C to Kelvin
C_to_K(44) # Converts from 44 C  to Kelvin
}
\seealso{
C_to_F, F_to_K
}
