% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_crop_ET.R
\docType{data}
\name{default_crop_ET}
\alias{default_crop_ET}
\title{default_crop_ET}
\format{
a data frame that contains a column for each crop type (wheat, corn, barley) and one observation:
\itemize{
\item ET values
}
}
\usage{
default_crop_ET
}
\description{
A data frame that stores crop type and respective ET values
}
\keyword{datasets}
