% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diamond_plot1.R
\name{diamond_plot1}
\alias{diamond_plot1}
\title{Diamond Plot}
\usage{
diamond_plot1(df, varx, vary)
}
\arguments{
\item{df}{a dataframe}

\item{varx}{a string name for variable x}

\item{vary}{a string name for variable y}
}
\value{
A plot with diamond markers of \code{vary} versus \code{varx}
}
\description{
This function produces a diamond plot
}
\examples{
2112
}
