% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_to_K.R
\name{F_to_K}
\alias{F_to_K}
\title{Fahrenheit to Kelvin}
\usage{
F_to_K(x)
}
\arguments{
\item{x}{Numeric vector of temperatures in Fahrenheit}
}
\value{
A numeric vector of temperatures in Kelvin
}
\description{
\code{F_to_K} converts the temperature units in Fahrenheit to Kelvin
}
\examples{
F_to_K(32) # Converts from 32 F to Kelvin
F_to_K(44) # Converts from 44 F  to Kelvin
}
\seealso{
C_to_F, C_to_K
}
