% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{function for printing hello}
\usage{
hello(word = "world")
}
\arguments{
\item{word}{that is printed after the hello}
}
\value{
string message
}
\description{
prints the hello string combined with the user input to the console
}
\examples{
hello()
hello("wim")

}
