% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irrigation.R
\name{irrigation}
\alias{irrigation}
\title{Irrigation Function}
\usage{
irrigation(precip, ET, crop_type)
}
\arguments{
\item{crop_type}{(wheat, corn, barley)}

\item{precipitation}{(precip) (mm/day)}

\item{evapotranspiration}{(ET) (mm/day)}
}
\value{
list of relative soil moisture availability, whether that crop should be irrigated, and how much water should be given (in L).
}
\description{
This function generates a list of irrigation recommendations:
(1) relative soil moisture availability (mm/day);
(2) whether crops should be irrigated (y/n); and,
(3) how much water should be given (mm/day)
based on three input parameters (precipitation, evapotranspiration, and crop type)
}
\author{
Eva Wilson, Nickolas McManus, Katheryn Moya
}
