% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/example.R
\name{MD}
\alias{MD}
\title{Documentation example for \code{maxygen}}
\usage{
MD(x)
}
\arguments{
\item{x}{Markdown is \strong{allowed} for parameters \emph{as well}.}
}
\value{
This is a \emph{return value}.
}
\description{
Note that you explicitly need to use \code{@title}, \code{@description} and
also \code{@details}.
}
\details{
We test each feature here.
}
\note{
Markdown can be used within a \emph{note}, \strong{of course}.
}
\section{Simple markup}{

This is \strong{strong} and there are \emph{both} \emph{emphasized}, probably as
italic.
}

\section{List}{

Bulleted list:
\itemize{
\item First item.
\item Second item.
}

Numbered list:
\enumerate{
\item Fist item.
\item Second item.
}

Nested lists:
\itemize{
\item Bulleted item containing
\enumerate{
\item a numbered
\item list.
}
\item Another bulleted item.
}

Nested lists of the same kind:
\itemize{
\item Bulleted first
\itemize{
\item containing
\item another bulleted
}
\item Bulleted second.
}
\enumerate{
\item Numbered first,
\enumerate{
\item containing
\item another numbered.
}
\item Numbered second.
}
}

\section{Code}{


Inline code is simple like \code{2 times 2}. Code blocks are also easy:\preformatted{markdown <- function(text, tags) {
  md <- markdown_xml(text, hardbreaks = TRUE)
  xml <- read_xml(md)
  trim(markdown_rparse(xml, tags))
}
}
}

\section{Links}{

This is just a link: \url{http://igraph.org}.
This is another link: \href{http://igraph.org}{igraph}.
}
\author{
This was written by \emph{Bugs Bunny}, so don't blame me.
}
\seealso{
Don't forget to read \emph{Writing R extensions}. Here is a link:
\url{https://cran.r-project.org/doc/manuals/r-release/R-exts.html}.

Other X.maxygen.examples.: \code{\link{MD2}}
}

