% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/example.R
\name{MD2}
\alias{MD2}
\title{Another example, to have a family}
\usage{
MD2(x)
}
\arguments{
\item{x}{This is ignored, sorry.}
}
\description{
Another example, to have a family
}
\seealso{
Other X.maxygen.examples.: \code{\link{MD}}
}

