% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precipitation.R
\docType{data}
\name{precipitation}
\alias{precipitation}
\title{Precipitation data for irrigation scheduling}
\format{
A string with 50 precipitation values where mean is 2.5 mm and sd is 1 mm. This is generated using random distribution.
}
\usage{
precipitation
}
\description{
A random distribution of precipitation data
}
\keyword{datasets}
