% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{read_data}
\usage{
read_data(file_name)
}
\arguments{
\item{file_name}{name of the file that contains file name}
}
\value{
a data frame with bike counts
}
\description{
here we describe what read_data does
}
