% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rustexample.R
\name{rust.mean}
\alias{rust.mean}
\title{Calculate the mean of a vector using Rust}
\usage{
rust.mean(v)
}
\arguments{
\item{v}{Vector to calculate the mean on}
}
\value{
Mean of the vector
}
\description{
Calculate the mean of a vector using Rust
}
\examples{
rust.mean(rnorm(64, mean=5, sd=1))
}
