% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rustexample.R
\name{rust.sort}
\alias{rust.sort}
\title{Sort a vector using Rust}
\usage{
rust.sort(v)
}
\arguments{
\item{v}{Vector to sort}
}
\value{
Sorted vector
}
\description{
Sort a vector using Rust
}
\examples{
rust.sort(c(5, 3, 1, 2, 4))
}
