% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_split_one.R
\name{str_split_one}
\alias{str_split_one}
\title{Split a String}
\usage{
str_split_one(string, pattern, n = Inf)
}
\arguments{
\item{x}{a character vector with one element}

\item{split}{what to split on}
}
\value{
a character vector
}
\description{
Split a String
}
\examples{
x <- "alpha,bravo,charlie,delta"
str_split_one(x, pattern = ",")
str_split_one(x, pattern = ",", n = 2)

y <- "192.168.0.1"
str_split_one(y, pattern = stringr::fixed("."))
}
