% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{strsplit1}
\alias{strsplit1}
\title{Split a String}
\usage{
strsplit1(x, split)
}
\arguments{
\item{x}{a character vector with one element}

\item{split}{what to split on}
}
\value{
a character vector
}
\description{
Split a String
}
\examples{
x <- "alpha,bravo,charlie,delta"
}
