% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test}
\alias{test}
\title{Test function}
\usage{
test()
}
\value{
A list of names of objects exposed in each module.
}
\description{
Does nothing but test that we can load modules from different virtual environments.
}
\examples{
test()
}
\author{
Aaron Lun
}
