% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_graph.R
\name{weighted_graph}
\alias{weighted_graph}
\title{weighted_graph

Graphical representation of weighted data set

This function calculates the weighted mean and plots it alongside data and
weight. It requires numeric vectors for values and weights and returns a
graphic.}
\usage{
weighted_graph(data, weight)
}
\arguments{
\item{data}{value vector}

\item{weight}{weights vector}
}
\value{
Graph for data and weight with weighted mean.
}
\description{
weighted_graph

Graphical representation of weighted data set

This function calculates the weighted mean and plots it alongside data and
weight. It requires numeric vectors for values and weights and returns a
graphic.
}
