% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/second_data.R
\docType{data}
\name{y}
\alias{y}
\title{Sample of 100 datapoints between 1 and 100}
\format{
numerical vextor of length 100
}
\source{
sample(100)
}
\usage{
y
}
\description{
A dataset to be used as example or value or weight
}
\keyword{datasets}
