% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{square}
\alias{square}
\title{Square a number}
\usage{
square(x)
}
\arguments{
\item{x}{A number to be squared}
}
\value{
A number
}
\description{
Computes the square of a number.
}
\examples{
square(10)
}
